+++ test/pybarch.py
@@ -1,4 +1,0 @@
-import barch
-
-k = barch.KeyMap()
-k.Set("a","va")